Reorganize Framework: clean up Touki-folder leftovers
BitConverters.cs (BitConverterExtensions) moves to Polyfills/System/ namespace System -- it polyfills BCL BitConverter members.

TypeExtensions.cs moves to Polyfills/System/ namespace System -- it polyfills Type.IsAssignableTo (BCL .NET 5+).

Touki/StringExtensions.CopyTo(Span<char>) (PR #107 polyfill of string.CopyTo from .NET Core 2.1+) merges into Polyfills/System/StringExtensions.cs as an extension(string source) block.

EnumExtensions stays in namespace Touki because it shares a partial class with internal EnumData consumed by Touki.Text.ValueStringBuilder; was briefly moved before the cross-namespace partial-class break was discovered.

NumberFormatInfoExtensions moves to Touki/Globalization/ folder (namespace Touki.Globalization) -- reflects into BCL internals so stays in Touki per user direction. Number.Formatting.cs adds 'using Touki.Globalization;' to follow the move.

Removes empty touki/Framework/System tree.
