+++ touki/Framework/Polyfills/System/Number.Formatting.cs
@@ -8,6 +8,7 @@
 using System.Buffers.Text;
 using System.Globalization;
 using Touki;
+using Touki.Globalization;
 using Touki.Resources;
 using Touki.Text;
