+++ touki/Framework/Polyfills/System/StringExtensions.cs
@@ -89,4 +89,33 @@ public static class StringExtensions
             return result;
         }
     }
+
+    extension(string source)
+    {
+        /// <summary>
+        ///  Copies the contents of this string into the destination span.
+        /// </summary>
+        /// <param name="destination">The span into which to copy this string's contents.</param>
+        /// <exception cref="ArgumentException">The destination span is shorter than the source string.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public unsafe void CopyTo(Span<char> destination)
+        {
+            if ((uint)source.Length <= (uint)destination.Length)
+            {
+                fixed (char* pSource = source)
+                fixed (char* pDestination = destination)
+                {
+                    Buffer.MemoryCopy(
+                        pSource,
+                        pDestination,
+                        destination.Length * sizeof(char),
+                        source.Length * sizeof(char));
+                }
+            }
+            else
+            {
+                ArgumentException.Throw(null, nameof(destination));
+            }
+        }
+    }
 }
