+++ touki/Framework/Touki/Globalization/NumberFormatInfoExtensions.cs
@@ -1,11 +1,11 @@
-﻿// Copyright (c) 2025 Jeremy W Kuhne
+// Copyright (c) 2025 Jeremy W Kuhne
 // SPDX-License-Identifier: MIT
 // See LICENSE file in the project root for full license information
 
 using System.Globalization;
 using System.Reflection;
 
-namespace Touki;
+namespace Touki.Globalization;
 
 internal static class NumberFormatInfoExtensions
 {
